TodoApp/requirements.txt
Django>=4.0,<5.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
